+++ test/bessely_test.jl
@@ -101,7 +101,7 @@ end
 @test bessely(0.5:1:10.5, 40.0) ≈ SpecialFunctions.bessely.(0.5:1:10.5, 40.0) rtol=1e-11
 
 # test Float16
-@test bessely(10, Float16(1.0)) isa Float16
+@test bessely(Int16(10), Float16(1.0)) isa Float16
 @test bessely(10.2f0, 1.0f0) isa Float32
 
 # test limits for small arguments see https://github.com/JuliaMath/Bessels.jl/issues/35
